codebeat badge
I'm a co-founder of codebeat. Is it fine for you to add codebeat badge to README file? It will link to report we generated for Motan.

codebeat is automated code review application that through a set of analyzers and tools helps developers to deliver code that is robust, secure, efficient and easy to maintain.

Your feedback would be a great help for us. It will help us calibrate codebeat. We released support for Java five weeks ago.

We aim high! We want to build a set of tool developers could trust and love.